requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.32
langchain-core==0.2.39
langfuse==2.60.0
python-dotenv==1.0.1
pytest==8.3.4
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
chromadb==0.5.5
langgraph-checkpoint-postgres==3.0.4
psycopg[binary]==3.3.3
psycopg-pool==3.3.0